changes requested by reviewers
