+++ rcsb/workflow/cli/BcifExec.py
@@ -24,38 +24,46 @@ logger = logging.getLogger(__name__)
 def main():
 
     parser = argparse.ArgumentParser(formatter_class=argparse.RawTextHelpFormatter)
-    subparsers = parser.add_subparsers(help="subcommand help")
-    splitParser = subparsers.add_parser("split", help="split list files")
-    splitParser.add_argument(
-        "--subtasks",
-        default=1,
-        required=False,
-        help="sublists for pdb list and csm list, if 0 will equal number of cpus",
-    )
-    splitParser.add_argument(
-        "--loadType",
-        default="incremental",
-        choices=["full", "incremental"],
-        required=False,
-    )
-    computeParser = subparsers.add_parser("compute", help="distribute computation")
-    computeParser.add_argument(
+    # settings
+    parser.add_argument(
         "--batch",
         default=0,
         required=False,
-        help="subdivisions of sublists with same rules as subtasks",
+        help="subdivisions of sublists for pdb list and csm list, if 0 will equal number of cpus",
     )
-    computeParser.add_argument(
+    parser.add_argument(
+        "--nfiles",
+        default=0,
+        required=True,
+        help="set 0 for all files, set less than N for a test run, will not produce exactly n files",
+    )
+    parser.add_argument(
         "--maxTempFiles",
         default=100,
         required=False,
         help="max reads before clear out temp files",
     )
-    computeParser.add_argument(
-        "--nfiles",
-        default=0,
-        required=True,
-        help="set 0 for all files, set less than N for a test run, will not produce exactly n files",
+    parser.add_argument(
+        "--outfileSuffix",
+        default=".bcif.gz",
+        required=False,
+        choices=[".bcif", ".bcif.gz"],
+        help="whether to use additional gzip compression",
+    )
+    # output folder structure, default none (save all output files in one folder)
+    parser.add_argument(
+        "--outputContentType",
+        action="store_true",
+        default=False,
+        required=False,
+        help="whether output paths should include a directory for the content type (pdb, csm)",
+    )
+    parser.add_argument(
+        "--outputHash",
+        action="store_true",
+        default=False,
+        required=False,
+        help="whether output paths should include the hash for the entry",
     )
     # paths
     parser.add_argument(
@@ -65,46 +73,19 @@ def main():
         help="output directory for bcif files",
     )
     parser.add_argument(
-        "--tempPath", default="/tmp", required=True, help="files are wiped on cleanup"
-    )
-    # not required (rely on defaults)
-    parser.add_argument(
-        "--outfileSuffix",
-        default=".bcif.gz",
-        required=False,
-        choices=[".bcif", ".bcif.gz"],
-        help="whether to use additional gzip compression",
-    )
-    parser.add_argument(
-        "--inputPath",
-        default="/mnt/vdb1/in",
-        required=False,
-        help="local workflow only",
-    )
-    parser.add_argument(
         "--listFileBase",
         default="/tmp",
         required=False,
         help="input lists, may be same as temp path",
     )
     parser.add_argument(
-        "--localInputsOrRemote",
-        default="remote",
-        choices=["local", "remote"],
-        required=False,
-    )
-    parser.add_argument("--statusStartFile", default="status.start", required=False)
-    parser.add_argument(
-        "--statusCompleteFile", default="status.complete", required=False
+        "--listFileName",
+        default="pdbx_core_ids-1.txt",
+        required=True,
+        help="name of list file to read",
     )
     parser.add_argument("--missingFileBase", default="/home/ubuntu", required=False)
-    parser.add_argument("--missingFileName", default="missing.txt", required=False)
-    parser.add_argument(
-        "--removedFileName",
-        default="removed.txt",
-        required=False,
-        help="saved to missingFileBase",
-    )
+    # config
     parser.add_argument(
         "--pdbxDict",
         default="https://raw.githubusercontent.com/wwpdb-dictionaries/mmcif_pdbx/master/dist/mmcif_pdbx_v5_next.dic",
@@ -145,37 +126,14 @@ def main():
     parser.add_argument(
         "--structureFilePath", default="data/structures/divided/mmCIF/", required=False
     )
-    parser.add_argument(
-        "--configPath",
-        default="./rcsb/workflow/bcif/config.yml",
-        required=False,
-        help="required for split id list",
-    )
-    # output folder structure, default none (save all output files in one folder)
-    parser.add_argument(
-        "--outputContentType",
-        action="store_true",
-        default=False,
-        required=False,
-        help="whether output paths should include a directory for the content type (pdb, csm)",
-    )
-    parser.add_argument(
-        "--outputHash",
-        action="store_true",
-        default=False,
-        required=False,
-        help="whether output paths should include the hash for the entry",
-    )
 
     args = parser.parse_args()
+
+    logger.info("bcif workflow parsed args")
 
     try:
         (BcifWorkflow(args))()
     except RuntimeError as e:
-        raise Exception(str(e)) from e
-    except ValueError as e:
-        raise Exception(str(e)) from e
-    except Exception as e:
         raise Exception(str(e)) from e
 
 
@@ -185,5 +143,6 @@ if __name__ == "__main__":
         main()
     except Exception as e:
         logger.exception(str(e))
+        raise Exception(str(e)) from e
     t2 = time.time() - t
     logger.info("completed in %.2f s", t2)
